lib: rename lib/internal/vm/Module.js to lower case
