+++ node.gyp
@@ -143,8 +143,8 @@
       'lib/internal/v8.js',
       'lib/internal/v8_prof_polyfill.js',
       'lib/internal/v8_prof_processor.js',
-      'lib/internal/vm/Module.js',
       'lib/internal/stream_base_commons.js',
+      'lib/internal/vm/module.js',
       'lib/internal/streams/lazy_transform.js',
       'lib/internal/streams/async_iterator.js',
       'lib/internal/streams/buffer_list.js',
